Testing the problem
I tested the problem with the images and it fully works now along with the links in here. I am commiting this change to change the title of the read me so it no longer reads the default "code refactor" but instead says read me for horiseon which is directed more towards the company that I intend to have them read.